Fixed noise with ext. input and added backwards dir
When chained to external input, the plug-in was making noise when the
input audio was silence. This is because with analog input the signal
sometimes variates very slightly from 0.0. This was amplified by the bit
manipulation and turned into noise.

Also added the possibility to switch backwards in the buttons colour
with right click.